requirements.txt
mysql-connector-python
flask
fastapi
uvicorn